+++ astropy/time/formats.py
@@ -97,10 +97,11 @@ class TimeFormat(metaclass=TimeFormatMeta):
 
     Parameters
     ----------
-    val1 : numpy ndarray, list, str, or number
-        Data to initialize table.
-    val2 : numpy ndarray, list, str, or number; optional
-        Data to initialize table.
+    val1 : numpy ndarray, list, number, str, or bytes
+        Values to initialize the time or times.  Bytes are decoded as ascii.
+    val2 : numpy ndarray, list, or number; optional
+        Value(s) to initialize the time or times.  Only used for numerical
+        input, to help preserve precision.
     scale : str
         Time scale of input value(s)
     precision : int
@@ -735,13 +736,15 @@ class TimeString(TimeUnique):
         """Parse the time strings contained in val1 and set jd1, jd2"""
         # Select subformats based on current self.in_subfmt
         subfmts = self._select_subfmts(self.in_subfmt)
-
+        # Be liberal in what we accept: convert bytes to ascii.
+        to_string = (str if val1.dtype.kind == 'U' else
+                     lambda x: str(x, encoding='ascii'))
         iterator = np.nditer([val1, None, None, None, None, None, None],
                              op_dtypes=[val1.dtype] + 5*[np.intc] + [np.double])
-
         for val, iy, im, id, ihr, imin, dsec in iterator:
+            val = to_string(val)
             iy[...], im[...], id[...], ihr[...], imin[...], dsec[...] = (
-                self.parse_string(val.item(), subfmts))
+                self.parse_string(val, subfmts))
 
         jd1, jd2 = erfa.dtf2d(self.scale.upper().encode('ascii'),
                               *iterator.operands[1:])
@@ -1057,15 +1060,18 @@ class TimeEpochDateString(TimeString):
 
     def set_jds(self, val1, val2):
         epoch_prefix = self.epoch_prefix
+        # Be liberal in what we accept: convert bytes to ascii.
+        to_string = (str if val1.dtype.kind == 'U' else
+                     lambda x: str(x, encoding='ascii'))
         iterator = np.nditer([val1, None], op_dtypes=[val1.dtype, np.double])
         for val, years in iterator:
-            time_str = val.item()
             try:
+                time_str = to_string(val)
                 epoch_type, year_str = time_str[0], time_str[1:]
                 year = float(year_str)
                 if epoch_type.upper() != epoch_prefix:
                     raise ValueError
-            except (IndexError, ValueError):
+            except (IndexError, ValueError, UnicodeEncodeError):
                 raise ValueError('Time {0} does not match {1} format'
                                  .format(time_str, self.name))
             else: