+++ CHANGES.rst
@@ -469,6 +469,9 @@ astropy.time
 - Initialization of Time instances now is consistent for all formats to
   ensure that ``-0.5 <= jd2 < 0.5``. [#6653]
 
+- Initialization of ``Time`` instances with ``Column`` is now possible
+  also if the ``Column`` has dtype ``S``. [#6823]
+
 astropy.units
 ^^^^^^^^^^^^^